+++ src/main/java/com/trego/dto/AddressDTO.java
@@ -1,9 +1,6 @@
 package com.trego.dto;
 
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class AddressDTO {
@@ -15,12 +12,12 @@ public class AddressDTO {
     private String mobileNo;
     private String name;
     private int addressType;
-    private BigDecimal lat;
-    private BigDecimal lng;
+    private Double lat;
+    private Double lng;
     private long userId;
 
 
-    public AddressDTO(Long id, String address, String city, String landmark, String pincode, BigDecimal lat, BigDecimal lng, long userId, String mobileNo, String name, int addressType) {
+    public AddressDTO(Long id, String address, String city, String landmark, String pincode, Double lat, Double lng, long userId, String mobileNo, String name, int addressType) {
         this.id = id;
         this.address = address;
         this.city = city;
