+++ src/main/java/com/trego/dto/BucketOrderRequestDTO.java
@@ -8,5 +8,5 @@ public class BucketOrderRequestDTO {
     private long addressId;
     private long preOrderId; // The original preorder ID
     private Long bucketId; // The selected bucket ID
-    private Long vendorId; // The vendor ID from the bucket
+    private Integer vendorId; // The vendor ID from the bucket
 }
