+++ src/main/java/com/trego/service/IMedicineService.java
@@ -11,9 +11,9 @@ public interface IMedicineService {
 
     MedicineDTO getMedicineById(Long id);
 
-    Page<MedicineWithStockAndVendorDTO> searchMedicines(String searchText, long vendorId, int page, int size);
+    Page<MedicineWithStockAndVendorDTO> searchMedicines(String searchText, Integer vendorId, int page, int size);
 
-    Page<MedicineWithStockAndVendorDTO> getMedicinesBySubcategory(Integer subcategoryId, int page, int size);
+    Page<MedicineWithStockAndVendorDTO> getMedicinesBySubcategory(Long subcategoryId, int page, int size);
 
 
 }
